finish legal notice
